Move breaking glob functionality to new paackage
